+++ Runtime/ARCollaborationDataBuilder.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections;
-using System.Runtime.InteropServices;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
 
@@ -24,38 +22,20 @@ namespace UnityEngine.XR.ARKit
         /// (with <see cref="Append(byte[], int)"/> or <see cref="Append(NativeArray{byte}, int)"/>)
         /// or it has already been <see cref="Dispose"/>d.
         /// </summary>
-        public bool hasData
-        {
-            get
-            {
-                return m_NSMutableData != IntPtr.Zero;
-            }
-        }
+        public bool hasData => m_NSMutableData.created;
 
         /// <summary>
         /// The number of bytes owned by this struct.
         /// </summary>
+        /// <seealso cref="Append(byte[])"/>
         /// <seealso cref="Append(byte[], int)"/>
-        /// <seealso cref="Append(NativeArray{byte}, int)"/>.
-        public int count
-        {
-            get
-            {
-                if (m_NSMutableData != IntPtr.Zero)
-                {
-                    return UnityARKit_NSMutableData_getLength(m_NSMutableData);
-                }
-                else
-                {
-                    return 0;
-                }
-            }
-        }
+        /// <seealso cref="Append(NativeSlice{byte}, int)"/>.
+        public int length => m_NSMutableData.created ? m_NSMutableData.length : 0;
 
         /// <summary>
         /// Converts the bytes accumulated through calls to Append to an <see cref="ARCollaborationData"/>.
         /// </summary>
-        /// <exception cref="System.NotSupportedException">Thrown if ARCollaborationData is not supported.
+        /// <exception cref="System.NotSupportedException">Thrown if <see cref="ARCollaborationData"/> is not supported.
         /// Check for support with <see cref="ARKitSessionSubsystem.supportsCollaboration"/>.</exception>
         /// <exception cref="System.InvalidOperationException">Thrown if <see cref="hasData"/> is false.</exception>
         public ARCollaborationData ToCollaborationData()
@@ -66,61 +46,88 @@ namespace UnityEngine.XR.ARKit
             if (!hasData)
                 throw new InvalidOperationException("No data to convert to ARCollaborationData.");
 
-            return new ARCollaborationData(UnityARKit_session_collaborationDataFromNSData(m_NSMutableData));
+            return new ARCollaborationData(m_NSMutableData.ToNSData());
         }
 
         /// <summary>
-        /// Appends <paramref name="bytes"/> to an existing array of bytes.
+        /// Appends <paramref name="size"/> <c>byte</c>s of the array <paramref name="buffer"/> to an existing array of bytes.
         /// </summary>
-        public unsafe void Append(byte[] bytes, int length)
+        /// <param name="buffer">A buffer containing <c>byte</c>s to append.</param>
+        /// <param name="offset">The offset within <paramref name="buffer"/> to start appending bytes to the internal array.</param>
+        /// <param name="size">The number of bytes from <paramref name="buffer"/> to append. Must be less than <c><paramref name="bytes"/>.Length + <paramref name="offset"/>></c>.</param>
+        /// <exception cref="System.ArgumentNullException">Thrown if <paramref name="buffer"/> is <c>null</c>.</exception>
+        /// <exception cref="System.ArgumentOutOfRangeException">Thrown if <paramref name="size"/> is less than zero.</exception>
+        /// <exception cref="System.ArgumentOutOfRangeException">Thrown if <paramref name="offset"/> is less than zero.</exception>
+        /// <exception cref="System.InvalidOperationException">Thrown if <paramref name="size"/> is less than zero or greater than the length of <paramref name="buffer"/>.</exception>
+        public unsafe void Append(byte[] buffer, int offset, int size)
         {
-            if (bytes == null)
-                throw new ArgumentNullException("bytes");
+            if (buffer == null)
+                throw new ArgumentNullException(nameof(buffer));
 
-            if (bytes.Length < length)
-                throw new ArgumentException("'length' cannot be greater than the size of the byte array.", "length");
+            if (size < 0)
+                throw new ArgumentOutOfRangeException(nameof(size), size, $"'{nameof(size)}' must be greater than or equal to zero.");
 
-            fixed (byte* ptr = bytes)
+            if (offset < 0)
+                throw new ArgumentOutOfRangeException(nameof(offset), offset, $"'{nameof(offset)}' must be greater than or equal to zero.");
+
+            if (buffer.Length < size + offset)
+                throw new InvalidOperationException($"Reading {size} bytes starting at offset {offset} would write past the end of '{nameof(buffer)}' (buffer length = {buffer.Length}).");
+
+            fixed (byte* ptr = &buffer[offset])
             {
-                Append(ptr, length);
+                AppendUnchecked(ptr, size);
+            }
+        }
+
+        /// <summary>
+        /// Appends all <c>byte</c>s of the array <paramref name="buffer"/> to an existing array of bytes.
+        /// </summary>
+        /// <param name="buffer">A buffer containing <c>byte</c>s to append.</param>
+        /// <exception cref="System.ArgumentNullException">Thrown if <paramref name="buffer"/> is <c>null</c>.</exception>
+        public unsafe void Append(byte[] buffer)
+        {
+            if (buffer == null)
+                throw new ArgumentNullException(nameof(buffer));
+
+            fixed (byte* ptr = buffer)
+            {
+                AppendUnchecked(ptr, buffer.Length);
             }
         }
 
         /// <summary>
         /// Appends <paramref name="bytes"/> to an existing array of bytes.
         /// </summary>
-        public unsafe void Append(NativeArray<byte> bytes, int length)
+        /// <param name="bytes">An array of bytes to append to the existing data.</param>
+        /// <exception cref="System.ArgumentException">Thrown if <paramref name="bytes"/> does not reference valid data.</exception>
+        public unsafe void Append(NativeSlice<byte> bytes)
         {
-            if (!bytes.IsCreated)
-                throw new ArgumentException("'bytes' has not been created, or has been disposed.", "bytes");
+            void* ptr = bytes.GetUnsafePtr();
+            if (ptr == null)
+                throw new ArgumentException("Invalid NativeSlice", nameof(bytes));
 
-            if (bytes.Length < length)
-                throw new ArgumentException("'length' cannot be greater than the size of the byte array.", "length");
-
-            Append((byte*)bytes.GetUnsafePtr(), bytes.Length);
+            AppendUnchecked(ptr, bytes.Length);
         }
 
         /// <summary>
         /// Releases the native resource.
         /// </summary>
-        public void Dispose()
-        {
-            UnityARKit_CFRelease(m_NSMutableData);
-            m_NSMutableData = IntPtr.Zero;
-        }
+        public void Dispose() => m_NSMutableData.Dispose();
 
         /// <summary>
         /// Appends data to the underlying NSMutableData array.
         /// </summary>
-        unsafe void Append(byte* bytes, int length)
+        /// <param name="bytes">A pointer to an array of bytes to append to the existing data.</param>
+        /// <param name="size">The number of bytes pointed to by <paramref name="bytes"/>.</param>
+        unsafe void AppendUnchecked(void* bytes, int size)
         {
-            if (m_NSMutableData == IntPtr.Zero)
+            if (m_NSMutableData.created)
             {
-                m_NSMutableData = UnityARKit_NSMutableData_createWithBytes(bytes, length);
+                m_NSMutableData.Append(bytes, size);
             }
             else
             {
-                UnityARKit_NSMutableData_append(m_NSMutableData, bytes, length);
+                m_NSMutableData = new NSMutableData(bytes, size);
             }
         }
 
@@ -128,36 +135,21 @@ namespace UnityEngine.XR.ARKit
         /// Computes a hash code suitable for use in a <c>Dictionary</c> or <c>HashSet</c>.
         /// </summary>
         /// <returns>A hash code suitable for use in a <c>Dictionary</c> or <c>HashSet</c>.</returns>
-        public override int GetHashCode()
-        {
-            unchecked
-            {
-                return m_NSMutableData.GetHashCode();
-            }
-        }
+        public override int GetHashCode() => m_NSMutableData.GetHashCode();
 
         /// <summary>
         /// IEquatable interface. Compares for equality.
         /// </summary>
         /// <param name="obj">The object to compare for equality.</param>
         /// <returns><c>true</c> if <paramref name="obj"/> is of type <see cref="ARCollaborationDataBuilder"/> and compares equal with <see cref="Equals(ARCollaborationDataBuilder)"/>.</returns>
-        public override bool Equals(object obj)
-        {
-            if (!(obj is ARCollaborationDataBuilder))
-                return false;
-
-            return Equals((ARCollaborationDataBuilder)obj);
-        }
+        public override bool Equals(object obj) => (obj is ARCollaborationDataBuilder) && Equals((ARCollaborationDataBuilder)obj);
 
         /// <summary>
         /// IEquatable interface. Compares for equality.
         /// </summary>
         /// <param name="other">The <see cref="ARCollaborationDataBuilder"/> to compare against.</param>
         /// <returns><c>true</c> if all fields of this <see cref="ARCollaborationDataBuilder"/> compare equal to <paramref name="other"/>.</returns>
-        public bool Equals(ARCollaborationDataBuilder other)
-        {
-            return (m_NSMutableData == other.m_NSMutableData);
-        }
+        public bool Equals(ARCollaborationDataBuilder other) => m_NSMutableData.Equals(other.m_NSMutableData);
 
         /// <summary>
         /// Compares for equality. Same as <see cref="Equals(ARCollaborationDataBuilder)"/>.
@@ -165,10 +157,7 @@ namespace UnityEngine.XR.ARKit
         /// <param name="lhs">The left-hand side of the comparison.</param>
         /// <param name="rhs">The right-hand side of the comparison.</param>
         /// <returns><c>true</c> if all fields of this <see cref="ARCollaborationDataBuilder"/> compare equal to <paramref name="other"/>.</returns>
-        public static bool operator ==(ARCollaborationDataBuilder lhs, ARCollaborationDataBuilder rhs)
-        {
-            return lhs.Equals(rhs);
-        }
+        public static bool operator ==(ARCollaborationDataBuilder lhs, ARCollaborationDataBuilder rhs) => lhs.Equals(rhs);
 
         /// <summary>
         /// Compares for inequality. Same as <c>!</c><see cref="Equals(ARCollaborationDataBuilder)"/>.
@@ -176,34 +165,8 @@ namespace UnityEngine.XR.ARKit
         /// <param name="lhs">The left-hand side of the comparison.</param>
         /// <param name="rhs">The right-hand side of the comparison.</param>
         /// <returns><c>true</c> if any of the fields of this <see cref="ARCollaborationDataBuilder"/> are not equal to <paramref name="other"/>.</returns>
-        public static bool operator !=(ARCollaborationDataBuilder lhs, ARCollaborationDataBuilder rhs)
-        {
-            return !lhs.Equals(rhs);
-        }
+        public static bool operator !=(ARCollaborationDataBuilder lhs, ARCollaborationDataBuilder rhs) => !lhs.Equals(rhs);
 
-        [DllImport("__Internal")]
-        static extern void UnityARKit_CFRelease(IntPtr ptr);
-
-        [DllImport("__Internal")]
-        static extern unsafe void UnityARKit_NSMutableData_append(
-            IntPtr nsdata,
-            byte* bytes,
-            int length);
-
-        [DllImport("__Internal")]
-        static extern unsafe IntPtr UnityARKit_NSMutableData_create();
-
-        [DllImport("__Internal")]
-        static extern unsafe IntPtr UnityARKit_NSMutableData_createWithBytes(
-            byte* bytes,
-            int length);
-
-        [DllImport("__Internal")]
-        static extern IntPtr UnityARKit_session_collaborationDataFromNSData(IntPtr nsdata);
-
-        [DllImport("__Internal")]
-        static extern int UnityARKit_NSMutableData_getLength(IntPtr nsdata);
-
-        internal IntPtr m_NSMutableData;
+        internal NSMutableData m_NSMutableData;
     }
 }