+++ Runtime/ARKitReferencePointSubsystem.cs
@@ -113,7 +113,11 @@ namespace UnityEngine.XR.ARKit
             static extern bool UnityARKit_refPoints_tryRemove(TrackableId referencePointId);
         }
 
+#if UNITY_2019_2_OR_NEWER
+        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.SubsystemRegistration)]
+#else
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
+#endif
         static void RegisterDescriptor()
         {
 #if UNITY_IOS && !UNITY_EDITOR