+++ Runtime/Depth/ARKitXRDepthSubsystem.cs
@@ -147,7 +147,11 @@ namespace UnityEngine.XR.ARKit
         }
 
         //this method is run on startup of the app to register this provider with XR Subsystem Manager
+#if UNITY_2019_2_OR_NEWER
+        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.SubsystemRegistration)]
+#else
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
+#endif
         static void RegisterDescriptor()
         {
 #if UNITY_IOS && !UNITY_EDITOR