+++ Runtime/PlaneTracking/ARKitXRPlaneSubsystem.cs
@@ -177,7 +177,11 @@ namespace UnityEngine.XR.ARKit
                 void* boundary);
         }
 
+#if UNITY_2019_2_OR_NEWER
+        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.SubsystemRegistration)]
+#else
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
+#endif
         static void RegisterDescriptor()
         {
 #if UNITY_IOS && !UNITY_EDITOR