+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.2.0-preview.2] - 2019-07-16
+### New
+- Add support for `NotTrackingReason`.
+- Add support for matching the ARCore framerate with the Unity one. See `XRSessionSubsystem.matchFrameRate`.
+- Expose the [priority](https://docs.unity3d.com/Packages/com.unity.xr.arkit@2.2/api/UnityEngine.XR.ARKit.ARCollaborationData.html#UnityEngine_XR_ARKit_ARCollaborationData_priority) property on the `ARCollaborationData`.
+- Add support for getting the ambient light intensity in lumens.
+
+### Fixes
+- Update ARKit 3 compatibility for Xcode 11 beta 3. This fixes
+  - [Collaborative sessions](https://docs.unity3d.com/Packages/com.unity.xr.arkit@2.2/api/UnityEngine.XR.ARKit.ARCollaborationData.html)
+  - Human body tracking
+
 ## [2.2.0-preview.1] - 2019-06-05
 - Adding support for ARKit 3 functionality: Human pose estimation, human segmentation images, session collaboration, multiple face tracking, and tracking a face (with front camera) while in world tracking (with rear camera).